Subdividable screens.
A terminal now is made of a stack of Screens, each of which is a set of
CharGrids. Prior to this commit, a terminal was just a stack of
CharGrids. Each Screen can be successively subdivided into smaller
grids, to divide the space of the screen up with different purposes.

This commit is lacking certain features, which is why it is being
developed on a separate branch for now. The lacking features are:

* Scrolling. As of this commit, no scrollback is supported in any
  configuration. In the long term, each individual grid will be able to
  scroll independently and retain its off-screen state.
* Splitting at certain points. Currently, only a grid can be split in
  half. A section of the screen already split cannot be split again.